+++ packages/expo-module-template/example/metro.config.js
@@ -1,4 +1,6 @@
 // Learn more https://docs.expo.io/guides/customizing-metro
+
+/** @type {import('expo/metro-config')} */
 const { getDefaultConfig } = require('expo/metro-config');
 const path = require('path');