+++ templates/expo-template-tabs/metro.config.js
@@ -1,7 +1,8 @@
 // Learn more https://docs.expo.io/guides/customizing-metro
+
+/** @type {import('expo/metro-config')} */
 const { getDefaultConfig } = require('expo/metro-config');
 
-/** @type {import('expo/metro-config').MetroConfig} */
 const config = getDefaultConfig(__dirname, {
   // [Web-only]: Enables CSS support in Metro.
   isCSSEnabled: true,